Update variable type

+++ src/main/java/bdv/bigcat/viewer/atlas/opendialog/DVIDResponse.java
@@ -39,9 +39,9 @@ public class DVIDResponse
 
 		public String[] VoxelUnits;
 
-		public long[] MinPoint;
+		public double[] MinPoint;
 
-		public long[] MaxPoint;
+		public double[] MaxPoint;
 
 		public long[] MinIndex;
